Put each `State` in its own module
