+++ src/main.rs
@@ -39,7 +39,11 @@ fn main() -> amethyst::Result<()> {
         );
 
     let assets_dir = app_root.join("assets");
-    let mut game = Application::new(assets_dir, pong::Welcome::default(), game_data)?;
+    let mut game = Application::new(
+        assets_dir,
+        pong::states::welcome::Welcome::default(),
+        game_data,
+    )?;
     game.run();
 
     Ok(())
